+++ CHANGELOG.md
@@ -4,6 +4,35 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-08-18
+
+### Added
+
+- search feature
+- increase view count when visit articles
+
+## [0.2.2] - 2020-08-18
+
+### Added
+
+- Locust test files for Rest APIs
+
+## [0.2.1] - 2020-08-17
+
+### Added
+
+- Test files for Rest APIs
+
+## [0.2.0] - 2020-08-14
+
+### Added
+
+- Related article feature
+
+### Changed
+
+- Refactor some code
+
 ## [0.1.0] - 2020-08-08
 
 ### Added